requirements.txt
google-generativeai
langchain-google-genai
flask
Pillow